back/requirements/base.txt
# ── Core Django ──────────────────────────────────────────────────────────────
Django>=5.0,<5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27             # OpenAPI schema generation

# ── Multi-Tenancy ─────────────────────────────────────────────────────────────
django-tenants>=3.6

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9

# ── Money / Currency ──────────────────────────────────────────────────────────
django-money>=3.4

# ── Async / WebSockets ────────────────────────────────────────────────────────
channels>=4.0
channels[daphne]>=4.0
channels-redis>=4.2

# ── Task Queue ────────────────────────────────────────────────────────────────
celery>=5.3
django-celery-beat>=2.5
redis>=5.0

# ── Caching ───────────────────────────────────────────────────────────────────
django-redis>=5.4

# ── Auth & Security ───────────────────────────────────────────────────────────
django-cors-headers>=4.3

# ── Filtering / Search ────────────────────────────────────────────────────────
django-filter>=23.5

# ── Audit / History ───────────────────────────────────────────────────────────
django-auditlog>=2.3
django-simple-history>=3.5

# ── File Storage ──────────────────────────────────────────────────────────────
django-storages[s3]>=1.14
boto3>=1.34

# ── Email ─────────────────────────────────────────────────────────────────────
django-anymail[sendgrid]>=10.2

# ── PDF Generation ────────────────────────────────────────────────────────────
WeasyPrint>=60.0

# ── Data Seeding ──────────────────────────────────────────────────────────────
pandas>=2.0
openpyxl>=3.1

# ── Configuration ─────────────────────────────────────────────────────────────
python-decouple>=3.8

# ── AI / LangChain ────────────────────────────────────────────────────────────
langchain>=0.2
langchain-openai>=0.1
openai>=1.0
